Crate movies

Structs

AttributeDefinition

Represents an attribute for describing the key schema for the table and indexes.

AwsErrorRetryPolicy

A retry policy that models AWS error codes as outlined in the SEP

Client

An ergonomic service client for DynamoDB_20120810.

Config

Service config.

DescribeTable

Operation shape for DescribeTable.

DescribeTableError

Error type for the DescribeTable operation.

DescribeTableInput

Represents the input of a DescribeTable operation.

DescribeTableOutput

Represents the output of a DescribeTable operation.

Duration

A Duration type to represent a span of time, typically used for system timeouts.

HashMap

A hash map implemented with quadratic probing and SIMD lookup.

KeySchemaElement

Represents a single element of a key schema. A key schema specifies the attributes

Operation

Opt

ProvisionedThroughput

Represents the provisioned throughput settings for a specified table or index. The settings

Query

Fluent builder constructing a request to Query.

Region

The region to send requests to.

RegionProviderChain

Load a region by selecting the first from a series of region providers.

SdkSuccess

Successful SDK Result

WaitForReadyTable

Hand-written waiter to retry every second until the table is out of Creating state

Enums

AttributeValue

Represents the data for an attribute.

Error

All possible error types for this service.

KeyType

RetryKind

RetryKind describes how a request MAY be retried for a given response

ScalarAttributeType

SdkError

Failed SDK Result

TableStatus

Value

Represents any valid JSON value.

Statics

PKG_VERSION

Crate version number.

Traits

ClassifyResponse

StructOpt

A struct that is converted from command line arguments.

Functions

create_table

main

A partial reimplementation of https://docs.amazonaws.cn/en_us/amazondynamodb/latest/developerguide/GettingStarted.Ruby.html in Rust

movies_in_year

parse_item

value_to_item

wait_for_ready_table

Construct a DescribeTable request with a policy to retry every second until the table is ready

Derive Macros

StructOpt

Generates the StructOpt impl.